cache: Make the SimpleCache group key generic
This allows arbitrary types to be used for group keys instead of just
strings. The main motivation behind this is being able to more easily
use structured data for the group identifier instead of having to encode
it as a string.

This simplifies (and even encourages) using grouping to differentiate
envoy instances, e.g. by what app it's deployed alongside.

Signed-off-by: Snow Pettersen <snowp@squareup.com>