cache: Make SimpleCache#watches thread safe
SimpleCache#watches publishes a non-thread safe member of SimpleCache
without synchronization which means that changes made on other threads
are not guaranteed to be visible.

This change synchronizes access to the watches method and makes a
defensive copy of the underlying data. To simplify this I changed the
datastructure from Map<Map<>> to Table, which let's me use
ImmutableTable.copyOf to copy the data more easily.

It also changes usages of #watches() internal to the class to reference
the field directly to avoid the cost of a copy.

Signed-off-by: Snow Pettersen <snowp@squareup.com>