cache/server: keep track of acked responses and respond to new hints
This makes the DiscoveryServer keep track of the latest acked response,
allowing it to respond to DiscoveryRequests that come in for the
currently acked version but with new resource hints. This reduces the
need for odering the xDs responses: if a CDS response adds a cluster but
we've already sent the EDS response for the current version, Envoy will
be able to retreive the latest version by updating the resource hint and
receiving a new response.

Signed-off-by: Snow Pettersen <snowp@squareup.com>