cache: add overload for getResourceName(Any) (#62)
In DiscoveryServer we use this method to get the resource names of acked
resources. The type of the passed in object is Any, so getResourceName
responds with "" since it doesn't match any of the xds resources. This
adds an overload which will unpack the Any message to the appropriate
Message before determining the resource name.

Signed-off-by: Snow Pettersen <snowp@squareup.com>