cache: Drop reactor, synchronously call watch's response handler (#65)
We weren't really doing much with reactor, so it just ended up adding needless complexity. This change drops usage of reactor and migrates to a more traditional response handler system. The main consequence is that we're now executing the handler in the main server thread, but the handler should not block on any long-running calls.

Signed-off-by: Joey Bratton <jbratton@salesforce.com>