Use the public maven central relase of com.google.api.grpc:proto-google-common-protos
This fixes an issue of duplicate and possibly incompatible classes for
com/google/api and com/google/rpc proto generated classes. If a project
includes both api and proto-google-common-protos dependencies, there
will be duplicate classes on the class path. This is because:
- these classes are vendored in this repo but not declared in the pom
- the version of proto-google-common-protos used is not a public release in maven central

Similar to the issue in #171 that #237 will fix.

Signed-off-by: Preston Bennes <preston.bennes@snowflake.com>